flask-tutorial/requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Dance==1.1.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
MarkupSafe==1.0
oauthlib==2.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
